The two rotating functions that we have (for strings and for arrays) are not optimal, and I didn't quite pick up on this in [`b3acd44`](https://github.com/Smaug123/ClassicalCiphers.jl/commit/b3acd440f120a86d8b1b898ec30521cef7383ada).  This commit optimises them:
```julia
julia> using Benchmarktools, Random

julia> A = rand(1000);

julia> @btime rotate_right_old(A, 3);
  4.059 μs (1 allocation: 7.94 KiB)

julia> @btime rotate_right_new(A, 3);
  622.573 ns (1 allocation: 7.94 KiB)

julia> rotate_right_old(A, 3) == rotate_right_new(A, 3)
true

julia> @btime rotate_left_old(A, 3);
  4.042 μs (1 allocation: 7.94 KiB)

julia> @btime rotate_left_new(A, 3);
  639.320 ns (1 allocation: 7.94 KiB)

julia> rotate_left_old(A, 3) == rotate_left_new(A, 3)
true

julia> S = randstring(1000);

julia> @btime rotate_left_str_old(S, 3);
  30.889 μs (13 allocations: 10.55 KiB)

julia> @btime rotate_left(S, 3);
  883.661 ns (3 allocations: 2.16 KiB)

julia> rotate_left_str_old(S, 3) == rotate_left(S, 3)
true

julia> @btime rotate_right_str_old(S, 3);
  30.821 μs (13 allocations: 10.55 KiB)

julia> @btime rotate_right(S, 3);
  885.314 ns (3 allocations: 2.16 KiB)

julia> rotate_right_str_old(S, 3) == rotate_right(S, 3)
true
```