@jakewilliami 2021-05-03
My bad, I need to add some tests for code coverage.  Will do that now.